+++ test/mix/tasks/phx.gen.storybook_test.exs
@@ -13,7 +13,7 @@ defmodule Mix.Tasks.Phx.Gen.StorybookTest do
   @tag :capture_log
   test "mix phx.gen.storybook generates a working storybook", config do
     in_tmp_project(config.test, fn ->
-      for _ <- 1..6, do: send(self(), {:mix_shell_input, :yes?, true})
+      for _ <- 1..8, do: send(self(), {:mix_shell_input, :yes?, true})
       Storybook.run([])
 
       [{index, _}] = Code.compile_file("storybook/_root.index.exs")
@@ -27,31 +27,52 @@ defmodule Mix.Tasks.Phx.Gen.StorybookTest do
       assert backend.config(:sandbox_class) == "phoenix-storybook-web"
 
       assert_file("assets/js/storybook.js")
+
       assert_file("assets/css/storybook.css", fn file ->
         assert String.contains?(file, ~s|@import "tailwindcss/base|)
       end)
 
-      assert_shell_receive :info, ~r|Starting storybook generation|
-      assert_shell_receive :info, ~r|creating lib/phoenix_storybook_web/storybook.ex|
-      assert_shell_receive :info, ~r|creating storybook/_root.index.exs|
-      assert_shell_receive :info, ~r|creating storybook/welcome.story.exs|
-      assert_shell_receive :info, ~r|creating storybook/core_components/button.story.exs|
-      assert_shell_receive :info, ~r|creating storybook/core_components/table.story.exs|
-      assert_shell_receive :info, ~r|creating assets/css/storybook.css|
-      assert_shell_receive :info, ~r|creating assets/js/storybook.js|
-      assert_shell_receive :yes?, ~r|Add the following to your.*router.ex.*:|
-      assert_shell_receive :yes?, ~r|Add.*js/storybook.js.*as a new entry point to your esbuild args in .*config/config.exs.*|
-      assert_shell_receive :yes?, ~r|Add a new Tailwind build profile for.*css/storybook.css.*in.*config/config.exs.*|
-      assert_shell_receive :yes?, ~r|Add a new.*endpoint watcher.*for your new Tailwind build profile in.*config/dev.exs.*|
-      assert_shell_receive :yes?, ~r|Add a new.*live_reload pattern.*to your endpoint in.*config/dev.exs.*|
-      assert_shell_receive :yes?, ~r|Add your storybook content to.*\.formatter.exs.*|
+      assert_shell_receive(:info, ~r|Starting storybook generation|)
+      assert_shell_receive(:info, ~r|creating lib/phoenix_storybook_web/storybook.ex|)
+      assert_shell_receive(:info, ~r|creating storybook/_root.index.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/welcome.story.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/core_components/button.story.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/core_components/table.story.exs|)
+      assert_shell_receive(:info, ~r|creating assets/css/storybook.css|)
+      assert_shell_receive(:info, ~r|creating assets/js/storybook.js|)
+      assert_shell_receive(:yes?, ~r|Add the following to your.*router.ex.*:|)
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add.*js/storybook.js.*as a new entry point to your esbuild args in .*config/config.exs.*|
+      )
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add a new Tailwind build profile for.*css/storybook.css.*in.*config/config.exs.*|
+      )
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add a new.*endpoint watcher.*for your new Tailwind build profile in.*config/dev.exs.*|
+      )
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add a new.*live_reload pattern.*to your endpoint in.*config/dev.exs.*|
+      )
+
+      assert_shell_receive(:yes?, ~r|Add your storybook content to.*\.formatter.exs.*|)
+      assert_shell_receive(:yes?, ~r|Add an alias to .*mix.exs.*|)
+      assert_shell_receive(:yes?, ~r|Add a COPY directive in .*Dockerfile.*|)
     end)
   end
 
   @tag :capture_log
-  test "mix phx.gen.storybook --no-tailwind generates a working storybook without tailwind", config do
+  test "mix phx.gen.storybook --no-tailwind generates a working storybook without tailwind",
+       config do
     in_tmp_project(config.test, fn ->
-      for _ <- 1..4, do: send(self(), {:mix_shell_input, :yes?, true})
+      for _ <- 1..5, do: send(self(), {:mix_shell_input, :yes?, true})
       Storybook.run(["--no-tailwind"])
 
       assert_file("storybook/_root.index.exs")
@@ -59,22 +80,33 @@ defmodule Mix.Tasks.Phx.Gen.StorybookTest do
       assert_file("lib/phoenix_storybook_web/storybook.ex")
 
       assert_file("assets/js/storybook.js")
+
       assert_file("assets/css/storybook.css", fn file ->
         refute String.contains?(file, ~s|@import "tailwindcss/base|)
       end)
 
-      assert_shell_receive :info, ~r|Starting storybook generation|
-      assert_shell_receive :info, ~r|creating lib/phoenix_storybook_web/storybook.ex|
-      assert_shell_receive :info, ~r|creating storybook/_root.index.exs|
-      assert_shell_receive :info, ~r|creating storybook/welcome.story.exs|
-      assert_shell_receive :info, ~r|creating storybook/core_components/button.story.exs|
-      assert_shell_receive :info, ~r|creating storybook/core_components/table.story.exs|
-      assert_shell_receive :info, ~r|creating assets/css/storybook.css|
-      assert_shell_receive :info, ~r|creating assets/js/storybook.js|
-      assert_shell_receive :yes?, ~r|Add the following to your.*router.ex.*:|
-      assert_shell_receive :yes?, ~r|Add.*js/storybook.js.*as a new entry point to your esbuild args in .*config/config.exs.*|
-      assert_shell_receive :yes?, ~r|Add a new.*live_reload pattern.*to your endpoint in.*config/dev.exs.*|
-      assert_shell_receive :yes?, ~r|Add your storybook content to.*\.formatter.exs.*|
+      assert_shell_receive(:info, ~r|Starting storybook generation|)
+      assert_shell_receive(:info, ~r|creating lib/phoenix_storybook_web/storybook.ex|)
+      assert_shell_receive(:info, ~r|creating storybook/_root.index.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/welcome.story.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/core_components/button.story.exs|)
+      assert_shell_receive(:info, ~r|creating storybook/core_components/table.story.exs|)
+      assert_shell_receive(:info, ~r|creating assets/css/storybook.css|)
+      assert_shell_receive(:info, ~r|creating assets/js/storybook.js|)
+      assert_shell_receive(:yes?, ~r|Add the following to your.*router.ex.*:|)
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add.*js/storybook.js.*as a new entry point to your esbuild args in .*config/config.exs.*|
+      )
+
+      assert_shell_receive(
+        :yes?,
+        ~r|Add a new.*live_reload pattern.*to your endpoint in.*config/dev.exs.*|
+      )
+
+      assert_shell_receive(:yes?, ~r|Add your storybook content to.*\.formatter.exs.*|)
+      assert_shell_receive(:yes?, ~r|Add a COPY directive in .*Dockerfile.*|)
     end)
   end
 
@@ -90,7 +122,6 @@ defmodule Mix.Tasks.Phx.Gen.StorybookTest do
     assert_receive {:mix_shell, ^kind, [msg]}
     assert msg =~ pattern
   end
-
 end
 
 defmodule PhoenixStorybookWeb.CoreComponents do