+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.5.2 (not released)
 
 - **bugfix**: [fixed issue with nested story modules](https://github.com/phenixdigital/phoenix_storybook/issues/260).
+- **improvement**: [improved mix phx.gen.storybook instructions](https://github.com/phenixdigital/phoenix_storybook/issues/252).
 
 ## v0.5.1 (2022-03-15)